revert

+++ codes/Model_v2.py
@@ -702,7 +702,7 @@ class KGEModel(nn.Module):
             smooth_targets = torch.zeros(batch_size, pred.size(1))
             for batch in range(batch_size):
                 targets[batch][positive_sample[batch][2]] = 1
-                smooth_targets[batch] = KGEModel.smooth_one_hot(targets[batch], pred.size(1), 0.1)
+                smooth_targets[batch] = KGEModel.smooth_one_hot(targets[batch].int(), pred.size(1), 0.1)
 
             if args.cuda:
                 pred = pred.cuda()
